Refactor beat sequencer into modular components
